Create Internet-Tweaks.reg
This is a mix of all the PC Tweaks I used on my computer. To sum this down to save time it does the following

Optimizes the LAN Servers to connect faster
makes the Packet Send Size the Largest It can be
Makes the Tick Frequency the lowest it can be
Sets the CongestionAlgorithm to be optimized with Lamans 
RecieveWindow To a optimized value 

And basically anything that is inside of the brackets inside of the file if you look at the code before you run and install this into your Registry Editor.